add support for custom IP addr source
add terraform argument "ip_attribute" which allows to change to source of the IP Address of the VM. The attribute is specified as a blank separated path within the <VM> tags of the opennebula vm info response.
the default value for "ip_attribute" is "TEMPLATE CONTEXT ETH0_IP"
add terraform argument "wait_for_attribute" which can be used to define a certain value within the <VM> tags of the opennebula vm info response that needs to be present for the vm creation to finish. This is helpful to ensure that the vm has completely started before any provisioners are executed, for example.
remove resourceVmRead parts which load vm info when no vm id is present. This should never be the case since the vm creation process heavily depends on the vm id to be present.