improve error handling when reading VM resources
-return the error when the API call to "one.vm.info" fails. Before, the
id was removed which would cause the resource to be removed from
terraforms point of view. This seems a little bit much, since the cause
of the error does not necessarily mean that the resource is no longer
available.
-return a custom error when the resource id is not available.
-do not set the resource id again while reading the resource. The id must
not change over the lifespan of the vm, since it is also the identifier
which is used for querying for the vm info. Therefore there is no point
in overriding the id during the read operation